optermized deleteUser

+++ src/main/java/org/example/notescollector/service/UserServiceImpl.java
@@ -7,6 +7,7 @@ import org.example.notescollector.dto.UserStatus;
 import org.example.notescollector.dto.impl.UserDTO;
 import org.example.notescollector.entity.impl.UserEntity;
 import org.example.notescollector.exception.DataPersistException;
+import org.example.notescollector.exception.UserNotFoundException;
 import org.example.notescollector.util.AppUtil;
 import org.example.notescollector.util.Mapping;
 import org.modelmapper.ModelMapper;
@@ -29,7 +30,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public void saveUser(UserDTO userDTO) {
         UserEntity saved = userDao.save(mapping.toUserEntity(userDTO));
-        if(saved == null) {
+        if (saved == null) {
             throw new DataPersistException("User not saved");
         }
 
@@ -42,23 +43,27 @@ public class UserServiceImpl implements UserService {
         List<UserEntity> all = userDao.findAll();
         return mapping.asUserDTOList(all);
 
-
     }
 
     @Override
-    public UserStatus getUser(String userId){
-      if(userDao.existsById(userId)){
-          UserEntity selectedUser = userDao.getReferenceById(userId);
-          return mapping.toUserDTO(selectedUser);
-      }else {
-          return new SelectedUserErrorStatus(2,"User not found");
-      }
+    public UserStatus getUser(String userId) {
+        if (userDao.existsById(userId)) {
+            UserEntity selectedUser = userDao.getReferenceById(userId);
+            return mapping.toUserDTO(selectedUser);
+        } else {
+            return new SelectedUserErrorStatus(2, "User not found");
+        }
 
     }
 
     @Override
     public void deleteUser(String userId) {
-        userDao.deleteById(userId);
+        Optional<UserEntity> existsById = userDao.findById(userId);
+        if (!existsById.isPresent()) {
+            throw new UserNotFoundException("User with id " + userId + " not found");
+        } else {
+            userDao.deleteById(userId);
+        }
     }
 
     @Override
